Update B3.py
fix the type of input

+++ B3.py
@@ -1,3 +1,3 @@
-a, b = map(int, input().split())
+a, b = map(float, input().split())
 
 print(a**3 + b**3 + a * b)
